Implemented logic for "insert or update"
SQLAlchemy does not come with a ready-to-use django style ".save()" method that dynamically either executes insert or update statements. So I implemented a preliminary version for Tweets. This could be made more general by building some sort of dependency tree which shows foreign key constraints and then uses this at every single node of the tree. Might be worth it for future refactorings.